Pamėgink savo ikelta atsisiust, kaip zip ir pamegink sukurti objektus ir juos kviesti main.py faile. Atkreipk demesi i pakeitimus ir kur jis susije, kas ka kviecia :)
