+++ abstract_shape/__init__.py
@@ -1,0 +1,6 @@
+from .abstract_shape import AbstractShape
+from .square import Square
+from .circle import Circle
+from .rectangle import Rectangle
+
+__all__ = ['Square', "AbstractShape", 'Circle', 'Rectangle']
