--- manoprojektas/__init__.py
@@ -1,6 +1,0 @@
-from .square import Square
-from .abstract_shape import AbstractShape
-from .circle import Circle
-from .rectangle import Rectangle
-
-__all__ = ['Square', "AbstractShape", 'Circle', 'Rectangle']
